cambiando de float a BigDecimal en Cuenta

+++ netbeans/bancoDominio/src/main/java/banco/bancodominio/Cuenta.java
@@ -4,6 +4,7 @@
  */
 package banco.bancodominio;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.Objects;
 
@@ -14,13 +15,13 @@ import java.util.Objects;
 public class Cuenta {
     private String num_cuenta;
     private String estado;
-    private float saldo;
+    private BigDecimal saldo;
     private Date fecha_creacion;
 
     public Cuenta() {
     }
 
-    public Cuenta(String num_cuenta, String estado, float saldo, Date fecha_creacion) {
+    public Cuenta(String num_cuenta, String estado, BigDecimal saldo, Date fecha_creacion) {
         this.num_cuenta = num_cuenta;
         this.estado = estado;
         this.saldo = saldo;
@@ -43,11 +44,11 @@ public class Cuenta {
         this.estado = estado;
     }
 
-    public float getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(float saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
